Split proto3 JSON writer and parser to separate files
