+++ protobuf/lib/protobuf.dart
@@ -38,7 +38,8 @@ part 'src/protobuf/json.dart';
 part 'src/protobuf/pb_list.dart';
 part 'src/protobuf/pb_map.dart';
 part 'src/protobuf/protobuf_enum.dart';
-part 'src/protobuf/proto3_json.dart';
+part 'src/protobuf/proto3_json_reader.dart';
+part 'src/protobuf/proto3_json_writer.dart';
 part 'src/protobuf/rpc_client.dart';
 part 'src/protobuf/unknown_field_set.dart';
 part 'src/protobuf/utils.dart';
